state-res: Move resolve snapshot tests at the root of the module
When looking at the module, it is nicer to see the tests directly rather
than having to look in a submodule. The macros are moved to a submodule
instead.
